Refactor FLP builder
Make it easier to build a FLP pipeline outside of the FLP reconciler.
This will be useful for NETOBSERV-627

- Create a new file flp_pipeline_builder.go that contains the
  pipeline building code, extracted from flp_common_objects.go
- Make some of these functions exported, for external usage
- Some API enhancements such as builder.NewGRPCPipeline for consumer
  friendlyness
